requirements.txt
# asn1crypto>=1.5.1
boto3>=1.26.162
botocore>=1.29.162
# cachetools>=5.3.1
# certifi>=2023.5.7
# cffi>=1.15.1
# charset-normalizer>=3.1.0
# cloudpickle>=2.2.1
# cryptography>=41.0.1
# db-dtypes>=1.1.1
# filelock>=3.12.2
# google-api-core>=2.11.1
google-auth>=2.21.0
google-cloud-bigquery>=3.11.3
# google-cloud-core>=2.3.2
# google-crc32c>=1.5.0
# google-resumable-media>=2.5.0
# googleapis-common-protos>=1.59.1
# greenlet>=2.0.2
# grpcio>=1.56.0
# grpcio-status>=1.56.0
# idna>=3.4
# jmespath>=1.0.1
# numpy>=1.25.0
# oscrypto>=1.3.0
# packaging>=23.1
pandas>=2.0.2
# proto-plus>=1.22.3
# protobuf>=4.23.3
# pyarrow>=12.0.1
# pyasn1>=0.5.0
# pyasn1-modules>=0.3.0
# pycparser>=2.21
# pycryptodomex>=3.18.0
# PyJWT>=2.7.0
# pyOpenSSL>=23.2.0
# python-dateutil>=2.8.2
# pytz>=2023.3
# requests>=2.31.0
# rsa>=4.9
# s3transfer>=0.6.1
# six>=1.16.0
snowflake-connector-python[pandas]>=3.0.4
# snowflake-snowpark-python>=1.5.1
# snowflake-sqlalchemy>=1.4.7
# SQLAlchemy>=2.0.17
# typing_extensions>=4.6.3
# tzdata>=2023.3
# urllib3>=1.26
